backend/src/ingestion/cleaners.py
from __future__ import annotations

from bs4 import BeautifulSoup


def sanitize_soup(html: str) -> BeautifulSoup:
    soup = BeautifulSoup(html, "html.parser")
    for tag in soup(["script", "style", "noscript"]):
        tag.decompose()
    return soup