backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workways-rag-backend"
version = "0.1.0"
description = "RAG FAQ backend for Workways support website"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "faiss-cpu>=1.8.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "numpy>=1.26.0",
  "openai>=1.51.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.0",
  "python-dotenv>=1.0.1",
  "tenacity>=9.0.0",
  "uvicorn>=0.31.0"
]

[project.optional-dependencies]
test = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "respx>=0.21.0"
]

[tool.pytest.ini_options]
addopts = "-q"